Split shell shared key TTL tests
